requirements.txt
pillow==10.3.0
typing==3.7.4.3
opencv-python==4.9.0.80
cairocffi==1.7.1
cffi==1.17.1
jmespath==1.0.1
numpy==1.26.4
opencv-stubs==0.0.10
pycairo==1.27.0
pycparser==2.22
PyGObject==3.50.0
python-dateutil==2.9.0.post0
setuptools==59.6.0
six==1.16.0
types-Pillow==10.2.0.20240822
progressbar2==4.5.0
numba==0.60.0